102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,17 +8,17 @@
  */
 int main(void)
 {
-	int fib1 = 1; /* First Fibonacci number */
-	int fib2 = 2; /* Second Fibonacci number */
-	int fib3; /* Variable to store the next Fibonacci number */
+	unsigned long long fib1 = 1; /* First Fibonacci number */
+	unsigned long long fib2 = 2; /* Second Fibonacci number */
+	unsigned long long fib3; /* Variable to store the next Fibonacci number */
 	int count; /* Counter variable */
 
-	printf("%d, %d", fib1, fib2); /* Print the first two Fibonacci numbers */
+	printf("%llu, %llu", fib1, fib2); /* Print the first two Fibonacci numbers */
 
 	for (count = 3; count <= 50; count++)
 	{
 		fib3 = fib1 + fib2;
-		printf(", %d", fib3);
+		printf(", %llu", fib3);
 		fib1 = fib2;
 		fib2 = fib3;
 	}
